Fixed location checking rest controller (/location)

+++ src/main/java/com/ndb/auction/dao/oracle/presale/PreSaleDao.java
@@ -121,7 +121,7 @@ public class PreSaleDao extends BaseOracleDao {
 		}, status);
     }
 
-    public int updateSold(int id, double amount) {
+    public int updateSold(int id, Long amount) {
         String sql = "UPDATE TBL_PRESALE SET SOLD=SOLD+? WHERE ID=?";
         return jdbcTemplate.update(sql, amount, id);
     }
