+++ src/main/java/com/ndb/auction/dao/oracle/transactions/stripe/StripeAuctionDao.java
@@ -23,7 +23,7 @@ public class StripeAuctionDao extends BaseOracleDao {
 		StripeAuctionTransaction m = new StripeAuctionTransaction();
 		m.setId(rs.getInt("ID"));
 		m.setUserId(rs.getInt("USER_ID"));
-		m.setAmount(rs.getDouble("AMOUNT"));
+		m.setAmount(rs.getLong("AMOUNT"));
 		m.setCreatedAt(rs.getTimestamp("CREATED_AT").getTime());
         m.setConfirmedAt(rs.getTimestamp("UPDATED_AT").getTime());
 		m.setStatus(rs.getBoolean("STATUS"));
