+++ src/main/java/com/ndb/auction/models/transactions/coinpayment/CoinpaymentWalletTransaction.java
@@ -9,10 +9,10 @@ import lombok.Setter;
 @NoArgsConstructor
 public class CoinpaymentWalletTransaction extends CoinpaymentTransaction {
     
-    public CoinpaymentWalletTransaction(int userId, Double amount, String coin, String network, Double cryptoAmount, String cryptoType) {
+    public CoinpaymentWalletTransaction(int userId, Long amount, String coin, String network, Double cryptoAmount, String cryptoType) {
         this.userId = userId;
         this.network = network;
-        this.amount = 0.0;
+        this.amount = 0L;
         this.cryptoType = cryptoType;
         this.cryptoAmount = cryptoAmount;
         this.coin = coin;
