+++ src/main/java/com/ndb/auction/service/payment/stripe/StripeWalletService.java
@@ -36,7 +36,7 @@ public class StripeWalletService extends StripeBaseService implements ITransacti
 		try {
 			if(m.getPaymentIntentId() == null) {
 				PaymentIntentCreateParams.Builder createParams = PaymentIntentCreateParams.builder()
-					.setAmount(m.getAmount().longValue())
+					.setAmount(m.getAmount())
 					.setCurrency("USD")	
 					.setConfirm(true)
 					.setPaymentMethod(m.getPaymentMethodId())
@@ -87,7 +87,7 @@ public class StripeWalletService extends StripeBaseService implements ITransacti
 		try {
 			if(m.getPaymentIntentId() == null) {
 				PaymentIntentCreateParams.Builder createParams = PaymentIntentCreateParams.builder()
-						.setAmount(m.getAmount().longValue())
+						.setAmount(m.getAmount())
 						.setCurrency("USD")
 						.setCustomer(customerId)
 						.setConfirm(true)
